GCP/PythonCloudFunction/requirements.txt
pip>=21.3.1
wheel>=0.37.1
rsa>=4.8
pyasn1>=0.4.8
six>=1.16.0
cachetools>=5.2.0
certifi>=2022.6.15
urllib3>=1.26.9
requests>=2.28.1
google-api-python-client>=2.52.0
idna>=3.3
setuptools>=60.2.0
google-api-core>=2.8.2
google-cloud-storage>=2.4.0
google-cloud-core>=2.3.1
google-cloud-firestore>=2.5.3
firebase-admin>=5.2.0